+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Don't hang indefinitely and provide nicer error messages when a connection
   can't be established.
+- Remove `Login` and `Registration` tabs and consolidate into one panel.
 
 ## 3.2.1 (2017-08-29)